requirements.txt
fastapi==0.110.2
FastAPI-SQLAlchemy==0.2.1
python-dotenv
black==24.4.2
SQLAlchemy== 2.0.29
uvicorn== 0.29.0
mysql-connector-python==8.3.0
pytest==8.1.1
alembic==1.13.1
cryptography==42.0.5
passlib==1.7.4